% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrivalBoardApi.R
\name{arrivalBoardApi}
\alias{arrivalBoardApi}
\title{Get an arrival board for a specific station}
\usage{
arrivalBoardApi(id, date, time)
}
\arguments{
\item{id}{character, the internal ID of the station, received from \code{locationNameApi}.}

\item{date}{character, the date of the arrival board in format \code{YYYY-MM-DD}. Must not be a past date.}

\item{time}{character, the time of the arrival board in format \code{HH:MM:SS}.}
}
\value{
A \code{list} containing the \code{path}, \code{response} and \code{content} of the \code{GET} request.
}
\description{
\code{arrivalBoardApi} returns an arrival board for a specific station, date and time.
}
\details{
\code{arrivalBoardApi()} uses the API key stored by \code{openbahn_auth}.
}
\examples{
\dontrun{
# Set your API key
openbahn_auth("YOUR_KEY_HERE")
# Get an arrival board for a specific station, date and time
arrivalBoardApi("008000240", date = Sys.Date() + 1, time = "12:00")
}
}
\references{
\url{http://data.deutschebahn.com/dataset/api-fahrplan}
}
\author{
Philipp Ottolinger
}
