% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journeyDetailApi.R
\name{journeyDetailApi}
\alias{journeyDetailApi}
\title{Get details of a specific journey}
\usage{
journeyDetailApi(reference_url)
}
\arguments{
\item{reference_url}{character, a reference url of a journey obtained using \code{departureBoardApi()} or \code{arrivalBoardApi()}.}
}
\value{
A \code{list} containing the \code{path}, \code{response} and \code{content} of the \code{GET} request.
}
\description{
\code{journeyDetailApi} returns detailed information on a journey found using \code{departureBoardApi()} or \code{arrivalBoardApi()}.
}
\details{
\code{journeyDetailApi()} uses the API key stored by \code{openbahn_auth}.
}
\references{
\url{http://data.deutschebahn.com/dataset/api-fahrplan}
}
\author{
Philipp Ottolinger
}
