% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationNameApi.R
\name{locationNameApi}
\alias{locationNameApi}
\title{Find a location}
\usage{
locationNameApi(query)
}
\arguments{
\item{query}{character, a city or train station to look for}
}
\value{
A \code{list} containing the \code{path}, \code{response} and \code{content} of the \code{GET} request.
}
\description{
\code{locationNameApi()} enables you to find locations and their respective IDs and coordinates known by the API.
}
\details{
\code{locationNameApi()} uses the API key stored by \code{openbahn_auth}.
}
\references{
\url{http://data.deutschebahn.com/dataset/api-fahrplan}
}
\author{
Philipp Ottolinger
}
