% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbahn-package.R
\docType{package}
\name{openbahn-package}
\alias{openbahn-package}
\title{openbahn: Query the Deutsche Bahn Timetable API}
\description{
This package provides functions to query the Deutsche Bahn timetable 
API to find locations, get departure or arrival boards and to get 
further details for specific journeys.
}
\details{
At the moment, only the timetable for German long distance traffic 
is supported.

The use this package an API key is required. An API key can be 
obtained by mailing to \code{DBOpenData@deutschebahn.com}.

To learn more about \code{openbahn}, start with the vignettes:
 \code{browseVignettes(package = "openbahn")}
}
