% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbahn_auth.R
\name{openbahn_auth}
\alias{openbahn_auth}
\title{Store the API key in an environment variable}
\usage{
openbahn_auth(key)
}
\arguments{
\item{key}{character, the API key provided by \url{http://data.deutschebahn.com/}}
}
\description{
\code{openbahn_auth()} stores the API key in an environment variable to make it accessible for all \code{openbahn} functions.
}
\details{
The API key is stored in the environment variable \code{OPENBAHN_KEY}. In addition, \code{openbahn_auth()} sets another environment variable (\code{OPENBAHN_USER_AGENT}) holding the user agent which is used for HTTP calls.
}
\examples{
\dontrun{
openbahn_auth("YOUR_API_KEY_HERE")
}
}
\author{
Philipp Ottolinger
}
