% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbahn_journeys.R
\name{openbahn_journeys}
\alias{openbahn_journeys}
\title{Get details of a specific journey}
\usage{
openbahn_journeys(reference_url)
}
\arguments{
\item{reference_url}{character, a reference url of a journey obtained using \code{departureBoardApi()} or \code{arrivalBoardApi()}.}
}
\value{
A \code{list} containing the \code{path}, \code{response} and \code{content} of the \code{GET} request.
}
\description{
\code{openbahn_journeys} returns detailed information on a journey found using \code{departureBoardApi()} or \code{arrivalBoardApi()}.
}
\details{
\code{openbahn_journeys()} uses the API key stored by \code{openbahn_auth}.
}
\examples{
\dontrun{
# Set your API key
openbahn_auth("YOUR_KEY_HERE")
# Get an arrival or departure board for a specific station, date and time
dep <- departureBoardApi("008000240", date = Sys.Date() + 1, time = "12:00")
# Get a reference URL for a specific train in the arrival or departure board
ref <- dep$content$DepartureBoard$Departure$JourneyDetailRef$ref[1]
# Get train details
openbahn_journeys(ref)
}
}
\references{
\url{http://data.deutschebahn.com/dataset/api-fahrplan}
}
\author{
Philipp Ottolinger
}
